CJPM-17498: moved TopNav from experimental to components
