+++ packages/visual-stack-docs/src/containers/Components/Docs/index.js
@@ -31,6 +31,7 @@ import PopoverDocs from './popover';
 import CollapsiblePanelDocs from './collapsiblepanel';
 import PaginationDocs from './pagination';
 import PercentSliderDocs from './percentslider';
+import TopNavParent from "./topnav";
 
 const routeComponentMap = {};
 
@@ -55,6 +56,7 @@ addComponentRoute('pageheader', 'Page Header', <PageHeaderDocs />);
 addComponentRoute('percentslider', 'Percent Slider', <PercentSliderDocs />);
 addComponentRoute('select', 'Select', <SelectDocs />);
 addComponentRoute('sidenav', 'SideNav', <SideNavDocs />);
+addComponentRoute('topnav', 'TopNav', <TopNavParent />);
 addComponentRoute('slidingpanel', 'Sliding Panel', <SlidingPanelDocs />);
 addComponentRoute(
   'loading-animation',
