+++ packages/visual-stack-docs/src/containers/Experimental/Docs/index.js
@@ -10,7 +10,6 @@ import TextDocs from './text';
 import BoxDocs from './box';
 import Tree from './tree';
 import TriStateCheckboxDocs from './tristatecheckbox';
-import TopNavParent from "./topnav";
 
 const experimentalRouteMap = {};
 
@@ -22,7 +21,6 @@ const addComponentRoute = (path, linkName, component) => {
 addComponentRoute('text', 'Text', <TextDocs />);
 addComponentRoute('box', 'Box', <BoxDocs />);
 addComponentRoute('tree', 'Tree', <Tree />);
-addComponentRoute('topnav', 'TopNav Header', <TopNavParent />);
 addComponentRoute(
   'tristatecheckbox',
   'TriStateCheckbox',
